Formatting: Normalize behavior in `wp_check_invalid_utf8()`
`wp_check_invalid_utf8()` has been dependant on the runtime
configuration of the system running it. For systems without the
necessary Unicode support it would simply not perform any validation of
the input. Worse, the use of `icon()` is not only system-dependant, but
it also fails to perform substitution when requested, returning `false`
instead of the substituted string as described in the function docblock.

In this patch a newer spec-compliant fallback is provided while the
function itself defers to `mb_scrub()`. Both of these implementations
are spec-compliant (to UTF-8) and perform “Maximal subpart replacement”
when encountering invalid byte sequences — an important aspect to
maintain secure and quality interop between systems.